Update tame.lua
Sets Tame success rate to Charm success rate until a better implementation is found.  Current Tame success rate is 5-10% on EM mobs and doesn't factor in INT, so is just generally busted.  This at least makes the move useful.  Previous implementation ignores INT calculations and is based solely on Charm success chance anyways, so this isn't much of a stretch.